Implement shell and create a new thread to calculate fibonacci
I use fib.s which is implement non-recursive method, and in Makefile,
I add fib.s in the $(TARGET) line for make qemu command.

In os.c, the main point is how to receive the char type in by user, and it
is implemented by use "USART_FLAG_RXNE". The shell can check if user is press
"enter", "backspace" or valid command, maybe more command like "remember
previous command" can be implemented.

The shell is implemented by writing a funtion, and the method of the whole
function has 3 steps:
1.Receive what user type-in until "enter" is pressed.
2.Token everything user type and check valid command "fib".
3.Use 2 while loop to repeat shell.

When the thread is created, the thread call the my_shell function, and the
funtion will take the privilege until the program end, so there will be some
bugs if there are multi-threads.